# AWS Compliance query performance > EC2